fmpq_poly set_coeff
